backend/requirements.txt
aiohttp==3.6.2
aiomysql==0.0.20
aiosqlite==0.15.0
arrow==0.16.0
async-timeout==3.0.1
attrs==20.2.0
binaryornot==0.4.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cookiecutter==1.7.2
cryptography==3.1
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
jinja2-time==0.2.0
MarkupSafe==1.1.1
marshmallow==3.8.0
multidict==4.7.6
numpy==1.19.4
pandas==1.1.4
poyo==0.5.0
pycparser==2.20
PyMySQL==0.9.2
PyPika==0.39.1
python-dateutil==2.8.1
python-dotenv==0.14.0
python-slugify==4.0.1
pytz==2020.4
requests==2.24.0
six==1.15.0
text-unidecode==1.3
tortoise-orm==0.16.15
tqdm==4.51.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
yarl==1.5.1
